[Refactor]: Refactor the MCP module & add sync for snapshot and zilliz cloud.
