+++ packages/mcp/src/config.ts
@@ -1,0 +1,173 @@
+import { envManager } from "@zilliz/code-context-core";
+
+export interface CodeContextMcpConfig {
+    name: string;
+    version: string;
+    // Embedding provider configuration
+    embeddingProvider: 'OpenAI' | 'VoyageAI' | 'Gemini' | 'Ollama';
+    embeddingModel: string;
+    // Provider-specific API keys
+    openaiApiKey?: string;
+    openaiBaseUrl?: string;
+    voyageaiApiKey?: string;
+    geminiApiKey?: string;
+    // Ollama configuration
+    ollamaModel?: string;
+    ollamaHost?: string;
+    // Vector database configuration
+    milvusAddress?: string; // Optional, can be auto-resolved from token
+    milvusToken?: string;
+}
+
+export interface CodebaseSnapshot {
+    indexedCodebases: string[];
+    indexingCodebases: string[];  // List of codebases currently being indexed
+    lastUpdated: string;
+}
+
+// Helper function to get default model for each provider
+export function getDefaultModelForProvider(provider: string): string {
+    switch (provider) {
+        case 'OpenAI':
+            return 'text-embedding-3-small';
+        case 'VoyageAI':
+            return 'voyage-code-3';
+        case 'Gemini':
+            return 'gemini-embedding-001';
+        case 'Ollama':
+            return 'nomic-embed-text';
+        default:
+            return 'text-embedding-3-small';
+    }
+}
+
+// Helper function to get embedding model with provider-specific environment variable priority
+export function getEmbeddingModelForProvider(provider: string): string {
+    switch (provider) {
+        case 'Ollama':
+            // For Ollama, prioritize OLLAMA_MODEL over EMBEDDING_MODEL
+            const ollamaModel = envManager.get('OLLAMA_MODEL') || envManager.get('EMBEDDING_MODEL') || getDefaultModelForProvider(provider);
+            console.log(`[DEBUG] 🎯 Ollama model selection: OLLAMA_MODEL=${envManager.get('OLLAMA_MODEL') || 'NOT SET'}, EMBEDDING_MODEL=${envManager.get('EMBEDDING_MODEL') || 'NOT SET'}, selected=${ollamaModel}`);
+            return ollamaModel;
+        case 'OpenAI':
+        case 'VoyageAI':
+        case 'Gemini':
+        default:
+            // For other providers, use EMBEDDING_MODEL or default
+            return envManager.get('EMBEDDING_MODEL') || getDefaultModelForProvider(provider);
+    }
+}
+
+export function createMcpConfig(): CodeContextMcpConfig {
+    // Debug: Print all environment variables related to CodeContext
+    console.log(`[DEBUG] 🔍 Environment Variables Debug:`);
+    console.log(`[DEBUG]   EMBEDDING_PROVIDER: ${envManager.get('EMBEDDING_PROVIDER') || 'NOT SET'}`);
+    console.log(`[DEBUG]   EMBEDDING_MODEL: ${envManager.get('EMBEDDING_MODEL') || 'NOT SET'}`);
+    console.log(`[DEBUG]   OLLAMA_MODEL: ${envManager.get('OLLAMA_MODEL') || 'NOT SET'}`);
+    console.log(`[DEBUG]   GEMINI_API_KEY: ${envManager.get('GEMINI_API_KEY') ? 'SET (length: ' + envManager.get('GEMINI_API_KEY')!.length + ')' : 'NOT SET'}`);
+    console.log(`[DEBUG]   OPENAI_API_KEY: ${envManager.get('OPENAI_API_KEY') ? 'SET (length: ' + envManager.get('OPENAI_API_KEY')!.length + ')' : 'NOT SET'}`);
+    console.log(`[DEBUG]   MILVUS_ADDRESS: ${envManager.get('MILVUS_ADDRESS') || 'NOT SET'}`);
+    console.log(`[DEBUG]   NODE_ENV: ${envManager.get('NODE_ENV') || 'NOT SET'}`);
+
+    const config: CodeContextMcpConfig = {
+        name: envManager.get('MCP_SERVER_NAME') || "CodeContext MCP Server",
+        version: envManager.get('MCP_SERVER_VERSION') || "1.0.0",
+        // Embedding provider configuration
+        embeddingProvider: (envManager.get('EMBEDDING_PROVIDER') as 'OpenAI' | 'VoyageAI' | 'Gemini' | 'Ollama') || 'OpenAI',
+        embeddingModel: getEmbeddingModelForProvider(envManager.get('EMBEDDING_PROVIDER') || 'OpenAI'),
+        // Provider-specific API keys
+        openaiApiKey: envManager.get('OPENAI_API_KEY'),
+        openaiBaseUrl: envManager.get('OPENAI_BASE_URL'),
+        voyageaiApiKey: envManager.get('VOYAGEAI_API_KEY'),
+        geminiApiKey: envManager.get('GEMINI_API_KEY'),
+        // Ollama configuration
+        ollamaModel: envManager.get('OLLAMA_MODEL'),
+        ollamaHost: envManager.get('OLLAMA_HOST'),
+        // Vector database configuration - address can be auto-resolved from token
+        milvusAddress: envManager.get('MILVUS_ADDRESS'), // Optional, can be resolved from token
+        milvusToken: envManager.get('MILVUS_TOKEN')
+    };
+
+    return config;
+}
+
+export function logConfigurationSummary(config: CodeContextMcpConfig): void {
+    // Log configuration summary before starting server
+    console.log(`[MCP] 🚀 Starting CodeContext MCP Server`);
+    console.log(`[MCP] Configuration Summary:`);
+    console.log(`[MCP]   Server: ${config.name} v${config.version}`);
+    console.log(`[MCP]   Embedding Provider: ${config.embeddingProvider}`);
+    console.log(`[MCP]   Embedding Model: ${config.embeddingModel}`);
+    console.log(`[MCP]   Milvus Address: ${config.milvusAddress || (config.milvusToken ? '[Auto-resolve from token]' : '[Not configured]')}`);
+
+    // Log provider-specific configuration without exposing sensitive data
+    switch (config.embeddingProvider) {
+        case 'OpenAI':
+            console.log(`[MCP]   OpenAI API Key: ${config.openaiApiKey ? '✅ Configured' : '❌ Missing'}`);
+            if (config.openaiBaseUrl) {
+                console.log(`[MCP]   OpenAI Base URL: ${config.openaiBaseUrl}`);
+            }
+            break;
+        case 'VoyageAI':
+            console.log(`[MCP]   VoyageAI API Key: ${config.voyageaiApiKey ? '✅ Configured' : '❌ Missing'}`);
+            break;
+        case 'Gemini':
+            console.log(`[MCP]   Gemini API Key: ${config.geminiApiKey ? '✅ Configured' : '❌ Missing'}`);
+            break;
+        case 'Ollama':
+            console.log(`[MCP]   Ollama Host: ${config.ollamaHost || 'http://127.0.0.1:11434'}`);
+            console.log(`[MCP]   Ollama Model: ${config.embeddingModel}`);
+            break;
+    }
+
+    console.log(`[MCP] 🔧 Initializing server components...`);
+}
+
+export function showHelpMessage(): void {
+    console.log(`
+CodeContext MCP Server
+
+Usage: npx @zilliz/code-context-mcp@latest [options]
+
+Options:
+  --help, -h                          Show this help message
+
+Environment Variables:
+  MCP_SERVER_NAME         Server name
+  MCP_SERVER_VERSION      Server version
+  
+  Embedding Provider Configuration:
+  EMBEDDING_PROVIDER      Embedding provider: OpenAI, VoyageAI, Gemini, Ollama (default: OpenAI)
+  EMBEDDING_MODEL         Embedding model name (auto-detected if not specified)
+  
+  Provider-specific API Keys:
+  OPENAI_API_KEY          OpenAI API key (required for OpenAI provider)
+  OPENAI_BASE_URL         OpenAI API base URL (optional, for custom endpoints)
+  VOYAGEAI_API_KEY        VoyageAI API key (required for VoyageAI provider)
+  GEMINI_API_KEY          Google AI API key (required for Gemini provider)
+  
+  Ollama Configuration:
+  OLLAMA_HOST             Ollama server host (default: http://127.0.0.1:11434)
+  OLLAMA_MODEL            Ollama model name (default: nomic-embed-text)
+  
+  Vector Database Configuration:
+  MILVUS_ADDRESS          Milvus address (optional, can be auto-resolved from token)
+  MILVUS_TOKEN            Milvus token (optional, used for authentication and address resolution)
+
+Examples:
+  # Start MCP server with OpenAI (default) and explicit Milvus address
+  OPENAI_API_KEY=sk-xxx MILVUS_ADDRESS=localhost:19530 npx @zilliz/code-context-mcp@latest
+  
+  # Start MCP server with OpenAI and auto-resolve Milvus address from token
+  OPENAI_API_KEY=sk-xxx MILVUS_TOKEN=your-zilliz-token npx @zilliz/code-context-mcp@latest
+  
+  # Start MCP server with VoyageAI
+  EMBEDDING_PROVIDER=VoyageAI VOYAGEAI_API_KEY=pa-xxx MILVUS_TOKEN=your-token npx @zilliz/code-context-mcp@latest
+  
+  # Start MCP server with Gemini
+  EMBEDDING_PROVIDER=Gemini GEMINI_API_KEY=xxx MILVUS_TOKEN=your-token npx @zilliz/code-context-mcp@latest
+  
+  # Start MCP server with Ollama
+  EMBEDDING_PROVIDER=Ollama EMBEDDING_MODEL=nomic-embed-text MILVUS_TOKEN=your-token npx @zilliz/code-context-mcp@latest
+        `);
+}
